Added ActionListener functionallity to the button.
Re-did the code to be able to detect when the button is pressed, and execute a task when it is. Currently only prints some text to the console. TODO add the functionality to add an audio file to the program. This will replace the output to the console and add the audio file to a variable to then be handled by the user.